+++ server/src/main/scala/com/twitter/server/util/TwitterStats.scala
@@ -1,9 +1,0 @@
-package com.twitter.server.util
-
-import com.twitter.finagle.stats.StatsReceiver
-
-object TwitterStats {
-
-  def register(statsReceiver: StatsReceiver): Unit = ()
-
-}